+++ components/Navigation/styles.js
@@ -14,7 +14,7 @@ const NavBar = styled.nav`
 
 const NavLink = styled(Link)`
   align-items: center;
-  background-color: #ffffff;
+  background-color: #fffae6;
   border: 1px solid rgba(0, 0, 0, 0.1);
   border-radius: 0.25rem;
   box-shadow: rgba(0, 0, 0, 0.02) 0 1px 3px 0;
