+++ pages/list-page.js
@@ -2,8 +2,6 @@ import Header from "../components/Header/Header";
 import List from "../components/List/List";
 
 export default function ListPage({ customerData }) {
-  console.log(customerData);
-
   return (
     <>
       <Header />
