+++ styles.js
@@ -10,6 +10,7 @@ export default createGlobalStyle`
   body {
     margin: 0;
     font-family: system-ui;
+    background-color: white;
   }
 `;
 const Invitation = styled.h1`
